+++ test/lab.js
@@ -3,21 +3,9 @@
 var Promise = Yaku = require("../src/yaku");
 var utils = require("../src/utils")
 
-        var gen = utils.async(function gen () {
-            return {
-                next: function () {
-                    return {
-                        done: false,
-                        value: Yaku.reject("err")
-                    };
-                },
-                "throw": function (err) {
-                    return {
-                        done: true,
-                        value: Yaku.reject(err)
-                    };
-                }
-            };
-        })
 
-return gen().catch(function (v) { return console.log(v); });
+Promise.resolve(10).then(() => {
+    Promise.enableLongStackTrace();
+}).then(() => {
+    throw 10
+});