+++ docs/changelog.md
@@ -3,6 +3,7 @@
   - add: `span` option to `retry`
   - **API CHANGE** `Observable`'s `emit` now renamed to `next`
   - add: `error` method to `Observable`
+  - fix: a long stack trace undefined bug
 
 - v0.14